fixed bugs in pagerank

+++ graphene/test/pagerank_push/util.h
@@ -2,7 +2,7 @@
 
 typedef long index_t;
 typedef unsigned int  vertex_t;
-typedef float sa_t;
+typedef double sa_t;
 typedef unsigned char bit_t;
 typedef unsigned int comp_t;
